Move to env_vars.py
