+++ src/python/pants/backend/go/subsystems/golang.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 import logging
 import os
 
-from pants.engine.environment import EnvironmentVars
+from pants.engine.env_vars import EnvironmentVars
 from pants.option.option_types import BoolOption, StrListOption, StrOption
 from pants.option.subsystem import Subsystem
 from pants.util.memo import memoized_method
