+++ src/python/pants/backend/go/util_rules/go_bootstrap_test.py
@@ -8,7 +8,7 @@ from typing import Mapping
 from pants.backend.go.util_rules.go_bootstrap import GoBootstrap, compatible_go_version
 from pants.backend.go.util_rules.go_bootstrap import rules as go_bootstrap_rules
 from pants.core.util_rules.asdf_test import fake_asdf_root, materialize_indices
-from pants.engine.environment import CompleteEnvironmentVars
+from pants.engine.env_vars import CompleteEnvironmentVars
 from pants.engine.rules import QueryRule
 from pants.testutil.rule_runner import RuleRunner
