+++ src/python/pants/init/engine_initializer.py
@@ -16,7 +16,7 @@ from pants.bsp.protocol import BSPHandlerMapping
 from pants.build_graph.build_configuration import BuildConfiguration
 from pants.core.util_rules import environments, system_binaries
 from pants.core.util_rules.environments import determine_bootstrap_environment
-from pants.engine import desktop, environment, fs, platform, process
+from pants.engine import desktop, env_vars, fs, platform, process
 from pants.engine.console import Console
 from pants.engine.environment import EnvironmentName
 from pants.engine.fs import PathGlobs, Snapshot, Workspace
@@ -262,7 +262,7 @@ class EngineInitializer:
                 *collect_rules(locals()),
                 *build_files.rules(),
                 *fs.rules(),
-                *environment.rules(),
+                *env_vars.rules(),
                 *desktop.rules(),
                 *git_rules(),
                 *graph.rules(),
